+++ step_definitions/given.js
@@ -1,4 +1,4 @@
-var helpers = require('./helpers');
+var lib = require('./lib');
 
 module.exports = function () {
   // TODO: Given steps
