+++ step_definitions/then.js
@@ -1,6 +1,6 @@
-var helpers = require('./helpers');
-var navigation = helpers.navigation;
-var form = helpers.form;
+var lib = require('./lib');
+var navigation = lib.navigation;
+var form = lib.form;
 
 module.exports = function () {
   this.Then(/^I should be at "([^"]*)"$/, function (url) {
